pyproject.toml
[project]
name = "pokemon-card-generator"
version = "0.1.0"
description = "Generate printable Pokemon cards from PokeAPI"
readme = "README.md"
authors = [
    { name = "User", email = "user@example.com" }
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "reportlab>=4.0.4",
    "click>=8.1.7",
    "rich>=13.5.2",
    "pydantic>=2.4.0",
    "aiohttp>=3.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
pokemon-cards = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
